Rate Us Popup
This is a popup that will trigger after a certain amount of days + A certain amount of launches which can both be changed easily in the AppRater class. I havent been able to test the days untill it prompts though I know the popup works correctly so it definitly should work.