Better handling of app lifecycle processes
unregisterReceiver (and by extension registerReceiver) is related to the realm being open and closed. Best practice is to unregister receivers during onPause() (the official examples do not do this, mind you), but to be certain I attempt to close the realm during all relevant shutdown procedures.

To facilitate the use of onPause, I also had to make sure the same logic present in onStart was present in onResume (onPause will flow into onResume when a user returns to the activity, and connection procedures are run during onStart. It may be a good idea to simply move this functionality from onStart rather than repeat it).

Related to issue #3 , needs further investigation