Responsive Layouts added
Responsive layouts added.
The layouts have been tested on:
600DP:
Nexus 5X
Pixel 2
680dp:
Pixel 3
720dp:
PIxel 3 XL
Pixel 3a
780dp:
Pixel 3a XL

Leaving the 600DP strings etc, on the mainpage, so if the layouts are looking off on one of your phones I can be notified as to which layout and which phone is wrong.